+++ CHANGELOG.md
@@ -8,12 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Deprecated
 
-- Item Search methods `get_all_items` and `get_all_items_as_dict` are now deprecated.
+- Item Search methods `get_all_items` and `get_all_items_as_dict` are now deprecated,
+  and may be removed as early as v0.5.0.
   These have been deprecated because they have the potential to perform a large number
   of requests to the server and instantiate a large number of objects in memory.
   To a user, this is only visible as a large delay in the method call and/or the
   exhaustion of all available memory. The iterator methods `get_items` or
-  `get_itemcollections` should be used instead. [#206](https://github.com/stac-utils/pystac-client/pull/206)
+  `get_item_collections` should be used instead. [#206](https://github.com/stac-utils/pystac-client/pull/206)
 
 ### Added
 
@@ -131,7 +132,7 @@ are in a single HTTP session, handle pagination and respects conformance
 - Update to use PySTAC 1.1.0
 - IO changed to use PySTAC's new StacIO base class. 
 - `Search.item_collections()` renamed to `Search.get_item_collections()`
-- `Search.item_collections()` renamed to `Search.get_items()`
+- `Search.item()` renamed to `Search.get_items()`
 - Conformance is checked by each individual function that requires a particular conformance
 - STAC API testing URLs changed to updated APIs
 - `ItemSearch.get_pages()` function moved to StacApiIO class for general use